exercise_4/requirements.txt
Flask
numpy
gunicorn
joblib 
scikit-learn==1.3.2 
pillow 
opencv-python
pytest